Fixes for running under Pyromaniac.
Need to disable the unaligned loads, as they're not supported in the
same way (and might raise unaligned exceptions on other systems).

Updated the gitignore and gitattributes files to display nicely and
ignore generated files.

Fixes to the terminal code appear to have been present as bugs for
some time - extra NULs being written, and the terminal size was not
being set properly. I guess I only really tested it in TelnetD.